+++ day15/part2.cpp
@@ -4,7 +4,7 @@
 using namespace std;
 
 
-int Part2::solve(const vector<Disc> &discs) {
+size_t Part2::solve(const vector<Disc> &discs) {
     vector<Disc> updatedDiscs(discs);
     updatedDiscs.emplace_back(updatedDiscs.size() + 1, 11, 0);
     return Part1::solve(updatedDiscs);
